store: Add trivial test to test the tests
